pyproject.toml
[project]
name = "pyrex"
version = "0.1.0"
description = "DataHub for SEC company filing data"
readme = "README.md"
requires-python = ">=3.12"

# Combined dependencies for Producer, Consumer, and Flink Analytics
dependencies = [
    "apache-flink==2.2.0",    # Core Analytics Engine
    "kafka-python>=2.0.2",    # For the Producer/Consumer
    "edgartools>=3.0.0",      # SEC/EDGAR Scraping
    "boto3>=1.34.0",          # MinIO/S3 Interaction
    "pandas>=2.2.0",          # Data Manipulation
    "pyarrow>=15.0.0",        # Parquet support for MinIO
    "psycopg2-binary>=2.9.9", # Postgres Connector
    "requests>=2.31.0",       # HTTP for SEC API
    "plotly", #for visualization
    "dash" #for dashboard

]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.uv]
managed = true